Cleanup circuit library tests (#4816)
* start moving library tests to separate folder

* add IQP


* add permutation


* add HLF


* add Diagonal


* add MCMT


* add GraphState

Mantas Čepulkovskis <60826723+mantcep@users.noreply.github.com>

* add GMS


* add QV


* rm old test_lib

* add init

* update file docstrings

* PEP 3120

* code reformat

* remove ddt when possible

* port global R tests


* port fourier checking circuit


* port grover op and quadratic form

* fix lint in graph state
